src: suggest filing an issue when the user manually restarts gopls
